+++ Assets/YooAsset/Editor/AssetBundleCollector/AssetBundleCollector.cs
@@ -152,11 +152,14 @@ namespace YooAsset.Editor
 			}
 
 			Dictionary<string, CollectAssetInfo> result = new Dictionary<string, CollectAssetInfo>(1000);
-			bool isRawAsset = PackRuleName == nameof(PackRawFile);
+
+			// 检测是否为原生资源打包规则
+			IPackRule packRuleInstance = AssetBundleCollectorSettingData.GetPackRuleInstance(PackRuleName);
+			bool isRawFilePackRule = packRuleInstance.IsRawFilePackRule();
 
 			// 检测原生资源包的收集器类型
-			if (isRawAsset && CollectorType != ECollectorType.MainAssetCollector)
-				throw new Exception($"The raw file must be set to {nameof(ECollectorType)}.{ECollectorType.MainAssetCollector} : {CollectPath}");
+			if (isRawFilePackRule && CollectorType != ECollectorType.MainAssetCollector)
+				throw new Exception($"The raw file pack rule must be set to {nameof(ECollectorType)}.{ECollectorType.MainAssetCollector} : {CollectPath}");
 
 			if (string.IsNullOrEmpty(CollectPath))
 				throw new Exception($"The collect path is null or empty in group : {group.GroupName}");
@@ -168,7 +171,7 @@ namespace YooAsset.Editor
 				string[] findAssets = EditorTools.FindAssets(EAssetSearchType.All, collectDirectory);
 				foreach (string assetPath in findAssets)
 				{
-					if (IsValidateAsset(assetPath) && IsCollectAsset(assetPath))
+					if (IsValidateAsset(assetPath, isRawFilePackRule) && IsCollectAsset(assetPath))
 					{
                         if (IsMultiPlatform)
                         {
@@ -186,7 +189,7 @@ namespace YooAsset.Editor
 
 						if (result.ContainsKey(assetPath) == false)
 						{
-							var collectAssetInfo = CreateCollectAssetInfo(command, group, assetPath, isRawAsset);
+							var collectAssetInfo = CreateCollectAssetInfo(command, group, assetPath, isRawFilePackRule);
 							result.Add(assetPath, collectAssetInfo);
 						}
 						else
@@ -199,9 +202,9 @@ namespace YooAsset.Editor
 			else
 			{
 				string assetPath = CollectPath;
-				if (IsValidateAsset(assetPath) && IsCollectAsset(assetPath))
+				if (IsValidateAsset(assetPath, isRawFilePackRule) && IsCollectAsset(assetPath))
 				{
-					var collectAssetInfo = CreateCollectAssetInfo(command, group, assetPath, isRawAsset);
+					var collectAssetInfo = CreateCollectAssetInfo(command, group, assetPath, isRawFilePackRule);
 					result.Add(assetPath, collectAssetInfo);
 				}
 				else
@@ -231,22 +234,22 @@ namespace YooAsset.Editor
 			return result.Values.ToList();
 		}
 
-		private CollectAssetInfo CreateCollectAssetInfo(CollectCommand command, AssetBundleCollectorGroup group, string assetPath, bool isRawAsset)
+		private CollectAssetInfo CreateCollectAssetInfo(CollectCommand command, AssetBundleCollectorGroup group, string assetPath, bool isRawFilePackRule)
 		{
 			string address = GetAddress(group, assetPath);
-			string bundleName = GetBundleName(group, assetPath);
+			string bundleName = GetBundleName(command, group, assetPath);
 			List<string> assetTags = GetAssetTags(group);
-			CollectAssetInfo collectAssetInfo = new CollectAssetInfo(CollectorType, bundleName, address, assetPath, assetTags, isRawAsset);
+			CollectAssetInfo collectAssetInfo = new CollectAssetInfo(CollectorType, bundleName, address, assetPath, isRawFilePackRule, assetTags);
 
 			// 注意：模拟构建模式下不需要收集依赖资源
 			if (command.BuildMode == EBuildMode.SimulateBuild)
 				collectAssetInfo.DependAssets = new List<string>();
 			else
-				collectAssetInfo.DependAssets = GetAllDependencies(assetPath);
+				collectAssetInfo.DependAssets = GetAllDependencies(assetPath, isRawFilePackRule);
 
 			return collectAssetInfo;
 		}
-		private bool IsValidateAsset(string assetPath)
+		private bool IsValidateAsset(string assetPath, bool isRawFilePackRule)
 		{
 			if (assetPath.StartsWith("Assets/") == false && assetPath.StartsWith("Packages/") == false)
 			{
@@ -263,10 +266,31 @@ namespace YooAsset.Editor
 			if (type == typeof(LightingDataAsset))
 				return false;
 
-			// 忽略Unity无法识别的无效文件
-			// 注意：只对非原生文件收集器处理
-			if(PackRuleName != nameof(PackRawFile))
+			// 检测原生文件是否合规
+			if (isRawFilePackRule)
 			{
+				string extension = StringUtility.RemoveFirstChar(System.IO.Path.GetExtension(assetPath));
+				if (extension == EAssetFileExtension.unity.ToString() || extension == EAssetFileExtension.prefab.ToString() ||
+					extension == EAssetFileExtension.mat.ToString() || extension == EAssetFileExtension.controller.ToString() ||
+					extension == EAssetFileExtension.fbx.ToString() || extension == EAssetFileExtension.anim.ToString() ||
+					extension == EAssetFileExtension.shader.ToString())
+				{
+					UnityEngine.Debug.LogWarning($"Raw file pack rule can not support file estension : {extension}");
+					return false;
+				}
+
+				// 注意：原生文件只支持无依赖关系的资源
+				string[] depends = AssetDatabase.GetDependencies(assetPath, true);
+				if (depends.Length != 1)
+				{
+					UnityEngine.Debug.LogWarning($"Raw file pack rule can not support estension : {extension}");
+					return false;
+				}
+			}
+			else
+			{
+				// 忽略Unity无法识别的无效文件
+				// 注意：只对非原生文件收集器处理
 				if (type == typeof(UnityEditor.DefaultAsset))
 				{
 					UnityEngine.Debug.LogWarning($"Cannot pack default asset : {assetPath}");
@@ -282,7 +306,7 @@ namespace YooAsset.Editor
 		}
 		private bool IsIgnoreFile(string fileExtension)
 		{
-			foreach (var extension in YooAssetSettings.IgnoreFileExtensions)
+			foreach (var extension in DefaultFilterRule.IgnoreFileExtensions)
 			{
 				if (extension == fileExtension)
 					return true;
@@ -308,19 +332,22 @@ namespace YooAsset.Editor
 			string adressValue = addressRuleInstance.GetAssetAddress(new AddressRuleData(assetPath, CollectPath, group.GroupName, Address, IsMultiPlatform));
 			return adressValue;
 		}
-		private string GetBundleName(AssetBundleCollectorGroup group, string assetPath)
+		private string GetBundleName(CollectCommand command, AssetBundleCollectorGroup group, string assetPath)
 		{
 			System.Type assetType = AssetDatabase.GetMainAssetTypeAtPath(assetPath);
 			if (assetType == typeof(UnityEngine.Shader) || assetType == typeof(UnityEngine.ShaderVariantCollection))
-				return EditorTools.GetRegularPath(YooAssetSettings.UnityShadersBundleName).ToLower();
-
-			// 根据规则设置获取资源包名称
-			IPackRule packRuleInstance = AssetBundleCollectorSettingData.GetPackRuleInstance(PackRuleName);
-			string bundleName = packRuleInstance.GetBundleName(new PackRuleData(assetPath, CollectPath, group.GroupName));
-			if(YooAssetSettingsData.Setting.RegularBundleName)
-				return EditorTools.GetRegularPath(bundleName).Replace('/', '_').Replace('.', '_').ToLower();
+			{
+				// 获取着色器打包规则结果
+				PackRuleResult packRuleResult = DefaultPackRule.CreateShadersPackRuleResult();
+				return packRuleResult.GetMainBundleName(command.PackageName, command.UniqueBundleName);
+			}
 			else
-				return EditorTools.GetRegularPath(bundleName).ToLower();
+			{
+				// 获取其它资源打包规则结果
+				IPackRule packRuleInstance = AssetBundleCollectorSettingData.GetPackRuleInstance(PackRuleName);
+				PackRuleResult packRuleResult = packRuleInstance.GetPackRuleResult(new PackRuleData(assetPath, CollectPath, group.GroupName));
+				return packRuleResult.GetMainBundleName(command.PackageName, command.UniqueBundleName);
+			}
 		}
 		private List<string> GetAssetTags(AssetBundleCollectorGroup group)
 		{
@@ -329,13 +356,13 @@ namespace YooAsset.Editor
 			tags.AddRange(temper);
 			return tags;
 		}
-		private List<string> GetAllDependencies(string mainAssetPath)
+		private List<string> GetAllDependencies(string mainAssetPath, bool isRawFilePackRule)
 		{
 			List<string> result = new List<string>();
 			string[] depends = AssetDatabase.GetDependencies(mainAssetPath, true);
 			foreach (string assetPath in depends)
 			{
-				if (IsValidateAsset(assetPath))
+				if (IsValidateAsset(assetPath, isRawFilePackRule))
 				{
 					// 注意：排除主资源对象
 					if (assetPath != mainAssetPath)