+++ Assets/YooAsset/Editor/AssetBundleCollector/CollectAssetInfo.cs
@@ -14,7 +14,7 @@ namespace YooAsset.Editor
 		/// 资源包名称
 		/// </summary>
 		public string BundleName { private set; get; }
-
+		
 		/// <summary>
 		/// 可寻址地址
 		/// </summary>
@@ -26,14 +26,14 @@ namespace YooAsset.Editor
 		public string AssetPath { private set; get; }
 
 		/// <summary>
-		/// 资源分类标签
-		/// </summary>
-		public List<string> AssetTags { private set; get; }
-
-		/// <summary>
 		/// 是否为原生资源
 		/// </summary>
 		public bool IsRawAsset { private set; get; }
+
+		/// <summary>
+		/// 资源分类标签
+		/// </summary>
+		public List<string> AssetTags { private set; get; }
 
 		/// <summary>
 		/// 依赖的资源列表
@@ -41,14 +41,14 @@ namespace YooAsset.Editor
 		public List<string> DependAssets = new List<string>();
 
 
-		public CollectAssetInfo(ECollectorType collectorType, string bundleName, string address, string assetPath, List<string> assetTags, bool isRawAsset)
+		public CollectAssetInfo(ECollectorType collectorType, string bundleName, string address, string assetPath, bool isRawAsset, List<string> assetTags)
 		{
 			CollectorType = collectorType;
 			BundleName = bundleName;
 			Address = address;
 			AssetPath = assetPath;
-			AssetTags = assetTags;
 			IsRawAsset = isRawAsset;
+			AssetTags = assetTags;		
 		}
 	}
 }