+++ Assets/YooAsset/Runtime/Settings/YooAssetSettings.cs
@@ -6,24 +6,9 @@ namespace YooAsset
 	internal class YooAssetSettings : ScriptableObject
 	{
 		/// <summary>
-		/// AssetBundle文件的后缀名
-		/// </summary>
-		public string AssetBundleFileVariant = "bundle";
-
-		/// <summary>
-		/// 原生文件的后缀名
-		/// </summary>
-		public string RawBundleFileVariant = "rawfile";
-
-		/// <summary>
 		/// 清单文件名称
 		/// </summary>
 		public string PatchManifestFileName = "PatchManifest";
-
-		/// <summary>
-		/// 资源包名正规化（移除路径分隔符）
-		/// </summary>
-		public bool RegularBundleName = true;
 
 
 		/// <summary>
@@ -41,7 +26,7 @@ namespace YooAsset
 		/// </summary>
 		public const string PatchManifestFileVersion = "1.4.0";
 
-		
+
 		/// <summary>
 		/// 缓存的数据文件名称
 		/// </summary>
@@ -64,19 +49,8 @@ namespace YooAsset
 		public const string ReportFileName = "BuildReport";
 
 		/// <summary>
-		/// Unity着色器资源包名称
-		/// </summary>
-		public const string UnityShadersBundleName = "unityshaders";
-
-		/// <summary>
 		/// 内置资源目录名称
 		/// </summary>
 		public const string StreamingAssetsBuildinFolder = "BuildinFiles";
-
-
-		/// <summary>
-		/// 忽略的文件类型
-		/// </summary>
-		public static readonly string[] IgnoreFileExtensions = { "", ".so", ".dll", ".cs", ".js", ".boo", ".meta", ".cginc", ".hlsl" };
 	}
 }